+++ src/app/redux/store.ts
@@ -1,21 +1,22 @@
 import { configureStore } from "@reduxjs/toolkit";
-import orderInputReducer from "./orderInputSlice";
-import pairSelectorReducer from "./pairSelectorSlice";
-import orderBookReducer from "./orderBookSlice";
-import priceChartReducer from "./priceChartSlice";
-import accountHistoryReducer from "./accountHistorySlice";
-import radixReducer from "./radixSlice";
+import { orderInputSlice } from "./orderInputSlice";
+import { pairSelectorSlice } from "./pairSelectorSlice";
+import { orderBookSlice } from "./orderBookSlice";
+import { priceChartSlice } from "./priceChartSlice";
+import { accountHistorySlice } from "./accountHistorySlice";
+import { radixSlice } from "./radixSlice";
 
 export const store = configureStore({
   reducer: {
-    radix: radixReducer,
-    pairSelector: pairSelectorReducer,
-    orderInput: orderInputReducer,
-    orderBook: orderBookReducer,
-    priceChart: priceChartReducer,
-    accountHistory: accountHistoryReducer,
+    radix: radixSlice.reducer,
+    pairSelector: pairSelectorSlice.reducer,
+    orderInput: orderInputSlice.reducer,
+    orderBook: orderBookSlice.reducer,
+    priceChart: priceChartSlice.reducer,
+    accountHistory: accountHistorySlice.reducer,
   },
 });
 
 export type RootState = ReturnType<typeof store.getState>;
+export type AppStore = typeof store;
 export type AppDispatch = typeof store.dispatch;
