+++ src/app/subscriptions.ts
@@ -1,0 +1,68 @@
+import { Subscription } from "rxjs";
+import {
+  DataRequestBuilder,
+  RadixDappToolkit,
+} from "@radixdlt/radix-dapp-toolkit";
+import * as adex from "alphadex-sdk-js";
+import { radixSlice, WalletData } from "./redux/radixSlice";
+import { fetchBalances } from "./redux/pairSelectorSlice";
+import { pairSelectorSlice } from "./redux/pairSelectorSlice";
+import { orderBookSlice } from "./redux/orderBookSlice";
+import { orderInputSlice } from "./redux/orderInputSlice";
+import { updateCandles } from "./redux/priceChartSlice";
+import { accountHistorySlice } from "./redux/accountHistorySlice";
+import { AppStore } from "./redux/store";
+
+export type RDT = ReturnType<typeof RadixDappToolkit>;
+
+let rdtInstance: null | RDT = null;
+export function getRdt() {
+  return rdtInstance;
+}
+function setRdt(rdt: RDT) {
+  rdtInstance = rdt;
+}
+
+let subs: Subscription[] = [];
+
+export function initializeSubscriptions(store: AppStore) {
+  rdtInstance = RadixDappToolkit({
+    dAppDefinitionAddress:
+      "account_tdx_d_12yhr42q497c46lr9vxkfsln9e9f3kjj8t5qexznpeq3ewxxqnhm527",
+    networkId: 13,
+  });
+  rdtInstance.walletApi.setRequestData(
+    DataRequestBuilder.accounts().exactly(1)
+  );
+  subs.push(
+    rdtInstance.walletApi.walletData$.subscribe((walletData: WalletData) => {
+      const data: WalletData = JSON.parse(JSON.stringify(walletData));
+      store.dispatch(radixSlice.actions.setWalletData(data));
+
+      // TODO: can we subscribe to balances from somewhere?
+      store.dispatch(fetchBalances());
+    })
+  );
+  setRdt(rdtInstance);
+  adex.init();
+  subs.push(
+    adex.clientState.stateChanged$.subscribe((newState) => {
+      const serializedState: adex.StaticState = JSON.parse(
+        JSON.stringify(newState)
+      );
+
+      store.dispatch(pairSelectorSlice.actions.updateAdex(serializedState));
+      store.dispatch(orderInputSlice.actions.updateAdex(serializedState));
+      store.dispatch(orderBookSlice.actions.updateAdex(serializedState));
+      store.dispatch(updateCandles(serializedState.currentPairCandlesList));
+      store.dispatch(accountHistorySlice.actions.updateAdex(serializedState));
+    })
+  );
+}
+
+export function unsubscribeAll() {
+  subs.forEach((sub) => {
+    sub.unsubscribe();
+  });
+  subs = [];
+}
